added basic authentication

+++ controllers/post.js
@@ -1,10 +1,12 @@
 const Post = require('../models/post');
+const User = require('../models/user');
 const { getUniqueSlug } = require('../utils/string');
 const { Op } = require("sequelize");
-const e = require('express');
+
 exports.all = async (req, res, next) => {
     try {
-        const posts = await Post.findAll();
+        const posts = await req.user.getPosts({ include: User });
+        // res.json({ posts });
         res.render('dashboard/posts', { title: 'News and Stories', posts });
     } catch (e) {
         console.log(e);
@@ -27,13 +29,12 @@ exports.save = async (req, res, next) => {
         const content = req.body.content;
         const image = req.file;
 
-        const post = await Post.create({
+        const post = await req.user.createPost({
             title,
             slug,
             image: `/${image.path}`,
             content
         });
-
         res.redirect('/dashboard/posts');
     } catch (e) {
         console.log(e);
@@ -44,7 +45,7 @@ exports.edit = async (req, res, next) => {
     try {
         const postId = req.params.postId;
         const post = await Post.findByPk(postId);
-        if (post) {
+        if (post && post.userId === req.user.id) {
             res.render('dashboard/posts/form', { title: 'News and Stories', post });
         } else {
             next(new Error('Post not found!'));
@@ -59,7 +60,7 @@ exports.update = async (req, res, next) => {
         const postId = req.params.postId;
         const post = await Post.findByPk(postId);
 
-        if (post) {
+        if (post && post.userId === req.user.id) {
             const title = req.body.title;
             const updateSlug = req.body.slug;
             const slug = updateSlug ? await getUniqueSlug(Post, title, updateSlug, post.id) : updateSlug;
@@ -91,7 +92,7 @@ exports.destroy = async (req, res, next) => {
     try {
         const post = await Post.findByPk(postId);
 
-        if (post) {
+        if (post && post.userId === req.user.id) {
             await Post.destroy({ where: { id: postId }});
             res.json({ message: 'Deleting post succeed!', status: 204 });
         } else {
